refactor: merge the results into the host screen
Merge the host and results screen into a single screen. This simplifies the navigation and makes
the screen a bit easier to use.
