+++ features/search/src/main/java/com/chesire/nekome/app/search/host/core/TrackSeriesUseCase.kt
@@ -1,4 +1,4 @@
-package com.chesire.nekome.app.search.results.core
+package com.chesire.nekome.app.search.host.core
 
 import com.chesire.nekome.core.Resource
 import com.chesire.nekome.core.flags.SeriesType
