--- features/search/src/main/res/navigation/search_nav_graph.xml
@@ -1,29 +1,0 @@
-<?xml version="1.0" encoding="utf-8"?>
-<navigation xmlns:android="http://schemas.android.com/apk/res/android"
-    xmlns:app="http://schemas.android.com/apk/res-auto"
-    xmlns:tools="http://schemas.android.com/tools"
-    android:id="@+id/search_nav_graph"
-    app:startDestination="@id/hostFragment"
-    tools:ignore="UnusedNavigation">
-    <!-- UnusedNavigation is used because this is reached as a nested graph -->
-
-    <fragment
-        android:id="@+id/hostFragment"
-        android:name="com.chesire.nekome.app.search.host.HostFragment"
-        android:label="@string/nav_search">
-        <action
-            android:id="@+id/toResultsFragment"
-            app:destination="@id/resultsFragment" />
-    </fragment>
-    <fragment
-        android:id="@+id/resultsFragment"
-        android:name="com.chesire.nekome.app.search.results.ResultsFragment"
-        android:label="{searchTerm}">
-        <argument
-            android:name="searchTerm"
-            app:argType="string" />
-        <argument
-            android:name="searchResults"
-            app:argType="com.chesire.nekome.app.search.domain.SearchModel[]" />
-    </fragment>
-</navigation>
