+++ features/search/src/test/java/com/chesire/nekome/app/search/host/core/TrackSeriesUseCaseTest.kt
@@ -1,6 +1,6 @@
 @file:OptIn(ExperimentalCoroutinesApi::class)
 
-package com.chesire.nekome.app.search.results.core
+package com.chesire.nekome.app.search.host.core
 
 import com.chesire.nekome.core.Resource
 import com.chesire.nekome.core.flags.SeriesStatus
